Extract editor tests to their own file
